+++ CHANGES.md
@@ -10,6 +10,7 @@
 ### Fixes
  * Fixed drifting of encoded index when parsing nested JSON with a nesting depth > 2.
  * Fixed `content` fields in raw JSON text not being mandatory.
+ * Fixed wrong color of fold margin in dark mode.
 
 
 ## 0.9.0-alpha